+++ db/postgres/db.js
@@ -2,7 +2,7 @@ const pgp = require('pg-promise')({
   capSQL: true,
 });
 
-const connection = 'postgres://localhost:5432/wegot';
+const connection = 'postgres://localhost:5432/wegotgeo';
 
 const db = pgp(connection);
