--- db/postgres/writedata.js
@@ -1,1 +1,0 @@
-const fs = require('fs');
