+++ package.json
@@ -2,7 +2,7 @@
   "name": "WeGot-Recommendations",
   "version": "",
   "description": "nearby recommendations",
-  "author": "Stacy Lin",
+  "author": "Stacy Lin && Eric Sandoval",
   "license": "MIT",
   "engines": {
     "node": ">=6.13.0"
@@ -11,10 +11,9 @@
     "server": "nodemon ./server/index.js",
     "react-dev": "webpack -d --watch",
     "fec-seed": "node seed.js",
-    "sdc-seed-mongo": "node ./db/mongo/seed-optimize-cpu.js",
-    "sdc-init-postgres": "psql -h localhost -d postgres -p 5432 -a -q  -f ./db/postgres/INIT.sql",
-    "sdc-seed-postgres": "node ./db/postgres/seed-pg.js",
-    "init-psql-geo": "psql -h localhost -d postgres -p 5432 -a -q  -f ./db/spatialTests/Postgres/init.sql",
+    "seed-mongo": "node ./db/Mongo/seed.js",
+    "init-postgres": "psql -h localhost -d postgres -p 5432 -a -q  -f ./db/Postgres/init.sql",
+    "seed-postgres": "node ./db/Postgres/seed.js",
     "test": "jest",
     "start": "node ./server/index.js",
     "dockertest": "npm run seed-database && npm start",
